Working changes and refactor to enums
