+++ lib/cborPeer.ts
@@ -1,9 +1,10 @@
-import { Peer, type SerializerMapping } from "./peer";
+import { Peer, } from "./peer";
 import { Cbor } from "./dataconnection/StreamConnection/Cbor";
+import { SerializerMapping, defaultSerializers } from "./optionInterfaces";
 
 export class CborPeer extends Peer {
 	override _serializers: SerializerMapping = {
-		Cbor,
+		...defaultSerializers,
 		default: Cbor,
 	};
 }
