+++ lib/enums.ts
@@ -71,10 +71,17 @@ export enum DataConnectionErrorType {
 }
 
 export enum SerializationType {
+	/// Buffered Connections
 	Binary = "binary",
 	BinaryUTF8 = "binary-utf8",
 	JSON = "json",
 	None = "raw",
+
+	/// Streaming Connections
+	CBOR = "cbor",
+	MsgPack = "msgpack",
+
+	Default = "default",
 }
 
 export enum SocketEventType {
