+++ lib/msgPackPeer.ts
@@ -1,9 +1,10 @@
-import { Peer, type SerializerMapping } from "./peer";
-import { MsgPack } from "./exports";
+import { Peer, } from "./peer";
+import { MsgPack} from "./exports";
+import { SerializerMapping, defaultSerializers } from "./optionInterfaces";
 
 export class MsgPackPeer extends Peer {
 	override _serializers: SerializerMapping = {
-		MsgPack,
+		...defaultSerializers,
 		default: MsgPack,
 	};
 }
